feat(SignalService): SignalService
